pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "genicam_simulator"
description = "Modern, extensible Python project management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
keywords = [
]
authors = [
  { name = "singemar"},
]
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "harvesters",
  "pyyaml",
  ]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/bobejo/genicam_simulator"

[project.scripts]
genicam_simulator = "genicam_simulator.cli:main"

[tool.hatch.version]
source = "vcs"


[tool.hatch.build.hooks.vcs]
version-file = "src/hatch/version.py"

[tool.hatch.build.targets.sdist]
exclude = []

[tool.hatch.envs.hatch-test]
default-args = ["tests"]

[tool.hatch.envs.hatch-static-analysis]
config-path = "none"

[tool.mypy]
disallow_untyped_defs = false
disallow_incomplete_defs = false
enable_error_code = ["ignore-without-code", "truthy-bool"]
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
warn_no_return = false
warn_unused_ignores = true
